Renamed Embed.js to ExtensioConfig.js
