+++ lib/components/Editor/CustomExtensions/useCustomExtensions.js
@@ -19,7 +19,7 @@ import Placeholder, {
   placeholderGenerator,
 } from "./Placeholder/ExtensionConfig";
 import Mention, { createMentionSuggestions } from "./Mention/ExtensionConfig";
-import Embeds from "./Embeds";
+import Embeds from "./Embeds/ExtensionConfig";
 import EmojiSuggestion from "./Emoji/EmojiSuggestion/ExtensionConfig";
 import EmojiPicker from "./Emoji/EmojiPicker/ExtensionConfig";
